+++ janetrs_version/src/lib.rs
@@ -1,6 +1,5 @@
 use core::{cmp::Ordering, fmt};
 
-use const_fn::const_fn;
 use evil_janet::{
     JANET_CURRENT_CONFIG_BITS, JANET_VERSION_MAJOR, JANET_VERSION_MINOR, JANET_VERSION_PATCH,
 };
@@ -41,7 +40,6 @@ impl JanetBuildConfig {
 
     /// Return `true` if Janet single threaded bit is set.
     #[inline]
-    #[const_fn("1.46")]
     pub const fn is_single_threaded(&self) -> bool {
         match self.bits {
             0 | 1 => false,
@@ -52,7 +50,6 @@ impl JanetBuildConfig {
 
     /// Return `true` is Janet nanbox bit is set.
     #[inline]
-    #[const_fn("1.46")]
     pub const fn is_nanbox(&self) -> bool {
         match self.bits {
             0 | 2 => false,