+++ src/client.rs
@@ -7,8 +7,6 @@ use core::{
 #[cfg(feature = "std")]
 use std::{error::Error as StdError, thread_local};
 
-use const_fn::const_fn;
-
 use crate::{
     env::{CFunOptions, DefOptions, JanetEnvironment, VarOptions},
     Janet, JanetTable,
@@ -336,7 +334,6 @@ impl JanetClient {
 
     /// Return a reference of the environment table of the runtime if it exist.
     #[inline]
-    #[const_fn("1.48")]
     pub const fn env(&self) -> Option<&JanetEnvironment> {
         self.env_table.as_ref()
     }